base: do not consume buffer timestamp for a 0-length output buffer
SN might return 0-length buffer typically as output buffer corresponding
to a codec-data input buffer (or alike).  In any case, it does not represent
real frame data and as such should not consume a real buffer's timestamp.
Moreover, if this occurs very early on as part of codec-data handling,
then there may not yet be any input timestamps stored (so count == 0),
which would temporarily then lead to count == -1.
On the other hand, if a real ts is consumed this way, following buffers will
have wrong timestamps, most notable following a seek.

Signed-off-by: Mark Nauwelaerts <mark.nauwelaerts@collabora.co.uk>
Signed-off-by: Felipe Contreras <felipe.contreras@nokia.com>