td_jpegdec: compute the correct structure size
Because of the memory alignment, the create_args structure in jpegdec reports
a size of 36, although its real size is 34 (1 uint32_t and 15 uint16_t).

As the DSP doesn't have that memory alignment, the given structure size is
wrong and returns an error when creating the node.

This patch calculates the correct size for the DSP.

Signed-off-by: Víctor Manuel Jáquez Leal <vjaquez@igalia.com>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>