base: don't flush last pinned buffer
The processing has ended, no need to flush the last buffer (we already
got the needed data).

According to Mark Nauwelaerts, this might not work in certain
situations; a flush happens exactly at the same time the EOS is being
pushed. However, the worst that could happen is that the flush will not
be executed, which might or not might not cause a glitch.

Signed-off-by: Felipe Contreras <felipe.contreras@nokia.com>