Change dmm_buffer size/len semantics a bit
Basically, 'len' is supposed to have the size of the data contained,
whereas 'size' is the size of the actual buffer (with padding).

The problem is that before 'size' didn't have any padding, so 'len' was
automatically filled right before sending the buffer based on this
'size', but now we don't have the unpadded 'size' any more, thanks to
f2b6efb[1].

We really need this padded size, so instead of using yet another field,
use 'len' which is unused from buffer allocation up to sending the
buffer.

This means that some code that used 'size' now should be using 'len'
instead.

Obviously this is a tricky change prone to trigger many issues, so
should be tested thorougly.

[1] dmm_buffer: store correct size (aligned)

Signed-off-by: Felipe Contreras <felipe.contreras@nokia.com>