base: fix access to correct dmm_buffer data
As Ивайло Димитров pointed out; the new size/len semantics cause the
'len' to be overwritten when the buffer is mapped, so we have to save it
to avoid that.

Independently, but also correctly, Mark Nauwelaerts found the same
problem whil accessing the 'need_copy' field, and probably 'data', so
essentially the whole dmm_buffer.

A simple way to fix the issues is to delay the map_buffer() call until
we have accessed all the fields we care about.

Signed-off-by: Felipe Contreras <felipe.contreras@nokia.com>